# perf: reduce mem allocs in internal/metrics